updated the update.sh script

+++ golang-install.sh
@@ -1,13 +1,26 @@
 #!/bin/sh
 
-VER=$(curl -s https://golang.org/VERSION?m=text)
+golang_ver=$(curl -s 'https://golang.org/VERSION?m=text')
 
-if [ ! -f "${VER}.linux-amd64.tar.gz" ]; then
-  wget -q "https://dl.google.com/go/${VER}.linux-amd64.tar.gz"
+
+if [ ! -x "$(command -v go)" ]; then
+  echo "golang needs to be installed"
+else
+  if grep -q "$golang_ver" <<< "$(go version)"; then
+    echo "golang is already up to date"
+  else
+    echo "updating golang"
+    exit 1
+    if [ ! -f "${golang_ver}.linux-amd64.tar.gz" ]; then
+      wget -q "https://dl.google.com/go/${golang_ver}.linux-amd64.tar.gz"
+    fi
+
+    tar xvf "${golang_ver}.linux-amd64.tar.gz"
+    sudo rm -rf /usr/local/go
+    sudo mv -v go /usr/local
+  fi
 fi
-tar xvf "${VER}.linux-amd64.tar.gz"
-sudo rm -rf /usr/local/go
-sudo mv -v go /usr/local
+
 go version
 
 exit 0
